# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/marks/verify": {
      "post": {
        "operationId": "MarkController_verifyTrue",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMarkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Marks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/marks/unverify": {
      "post": {
        "operationId": "MarkController_verifyFalse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMarkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Marks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/marks/create": {
      "post": {
        "operationId": "MarkController_createMark",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMarkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Marks"
        ]
      }
    },
    "/api/marks/admin/all": {
      "get": {
        "operationId": "MarkController_getAllMarks",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Marks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/marks/admin/{id}": {
      "delete": {
        "operationId": "MarkController_deleteMark",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Marks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/marks/admin/reindex": {
      "put": {
        "operationId": "MarkController_reindexSearhchEngine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Marks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/categories": {
      "get": {
        "operationId": "CategoriesController_getCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_deleteCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "CategoriesController_updateCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/categories/create": {
      "post": {
        "operationId": "CategoriesController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/categories/stats": {
      "get": {
        "operationId": "CategoriesController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/categories/search": {
      "get": {
        "operationId": "CategoriesController_search",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/categories/clear-cache": {
      "get": {
        "operationId": "CategoriesController_clearCache",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/categories/admin/reindex": {
      "put": {
        "operationId": "CategoriesController_reindexSearhchEngine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Incidents API",
    "description": "",
    "version": "0.1.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "type": "http"
      }
    },
    "schemas": {
...
      "VerifyMarkDto": {
        "type": "object",
        "properties": {
          "markId": {
            "type": "number",
            "description": "Уникальный идентификатор метки, которую необходимо проверить",
            "example": 123
          },
          "userId": {
            "type": "string",
            "description": "Уникальный идентификатор пользователя, который выполняет проверку",
            "example": "f7f4b6c9-95be-45b9-bc24-1324f9dbecfd"
          }
        },
        "required": [
          "markId",
          "userId"
        ]
      },
      "AddressDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Более широкий адрес",
            "example": "Страна, Город"
          },
          "name": {
            "type": "string",
            "description": "Адрес",
            "example": "Улица, дом 12"
          }
        },
        "required": [
          "description",
          "name"
        ]
      },
      "CreateMarkDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "Уникальный идентификатор пользователя, создающего метку",
            "example": "f7f4b6c9-95be-45b9-bc24-1324f9dbecfd"
          },
          "lat": {
            "type": "number",
            "description": "Широта места метки",
            "example": 52.52
          },
          "lng": {
            "type": "number",
            "description": "Долгота места метки",
            "example": 13.405
          },
          "title": {
            "type": "string",
            "description": "Название метки",
            "example": "Моя первая метка"
          },
          "description": {
            "type": "string",
            "description": "Описание метки",
            "example": "Это описание моей метки, которое может быть очень длинным, но не обязательным."
          },
          "categoryId": {
            "type": "number",
            "description": "Идентификатор категории метки",
            "example": 1
          },
          "address": {
            "description": "Адрес, связанный с меткой",
            "allOf": [
              {
                "$ref": "#/components/schemas/AddressDto"
              }
            ]
          }
        },
        "required": [
          "userId",
          "lat",
          "lng",
          "title",
          "categoryId"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название категории. Это обязательное поле.",
            "example": "Еда"
          },
          "color": {
            "type": "string",
            "description": "Цвет категории (необязательное поле). Может быть использовано для визуального выделения категории.",
            "example": "blue"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Уникальный идентификатор категории. Это обязательное поле.",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Новое название категории. Это обязательное поле.",
            "example": "Напитки"
          },
          "color": {
            "type": "string",
            "description": "Новый цвет категории (необязательное поле). Может быть использовано для визуального выделения.",
            "example": "green"
          }
        },
        "required": [
          "id",
          "name"
        ]
      }
    }
  }
}